services/api/requirements.txt
Flask==0.12.1
Flask-SQLAlchemy==2.1
Flask-Migrate==2.0.1
Flask-Restless==0.17.0
Jinja2==2.8
SQLAlchemy==1.1.9
Werkzeug==0.11.3
elasticsearch==5.3.0
gunicorn==19.7.1
marshmallow==2.13.5
pika==0.10.0
python-dateutil==2.4.2
urllib3==1.14
webargs==1.6.1
pyodbc==3.0.7
requests==2.13.0
alembic==0.9.1
simplejson==3.10.0
pymongo==3.4.0
bcrypt==3.1.3
